changing docstring, function name, files
